Adicionando mais modelos ao projeto

+++ graduacao/src/main/java/br/usp/exemplo/graduacao/servicos/CursoServico.java
@@ -1,33 +1,51 @@
 package br.usp.exemplo.graduacao.servicos;
 
-import br.usp.exemplo.graduacao.entidades.Curso;
-import br.usp.exemplo.graduacao.repositorios.CursoRepositorio;
-import org.springframework.stereotype.Service;
-
 import java.util.List;
-import java.util.Optional;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
+
+import br.usp.exemplo.graduacao.dtos.CursoDTO;
+import br.usp.exemplo.graduacao.entidades.Curso;
+import br.usp.exemplo.graduacao.forms.CursoForm;
+import br.usp.exemplo.graduacao.repositorios.CursoRepositorio;
+import lombok.RequiredArgsConstructor;
 
 @Service
+@RequiredArgsConstructor
 public class CursoServico {
     private final CursoRepositorio cursoRepositorio;
 
-    public CursoServico(CursoRepositorio cursoRepositorio) {
-        this.cursoRepositorio = cursoRepositorio;
+    public List<CursoDTO> listarTodos() {
+        return cursoRepositorio.findAll().stream().map(CursoDTO::new).toList();
     }
 
-    public List<Curso> listarTodos() {
-        return cursoRepositorio.findAll();
+    public CursoDTO buscarPorId(Long id) {
+        return cursoRepositorio.findById(id).map(CursoDTO::new).orElseThrow(
+                () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Curso não encontrado com o ID: " + id));
     }
 
-    public Optional<Curso> buscarPorId(int id) {
-        return cursoRepositorio.findById(id);
+    public CursoDTO salvar(CursoForm curso) {
+        Curso entidade = new Curso(curso);
+        Curso novo = this.cursoRepositorio.save(entidade);
+        return new CursoDTO(novo);
     }
 
-    public Curso salvar(Curso curso) {
-        return cursoRepositorio.save(curso);
+    public CursoDTO atualizar(Long id, CursoForm curso) {
+        Curso entidade = this.cursoRepositorio.findById(id).orElseThrow(
+                () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Curso não encontrado com o ID: " + id));
+        entidade.setNome(curso.getNome());
+        entidade.setDescricao(curso.getDescricao());
+        entidade.setSigla(curso.getSigla());
+        Curso retorno = this.cursoRepositorio.save(entidade);
+        return new CursoDTO(retorno);
     }
 
-    public void deletar(int id) {
+    public void deletar(Long id) {
+        if(!cursoRepositorio.existsById(id)){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Curso não encontrado com o ID: " + id);
+        }
         cursoRepositorio.deleteById(id);
     }
 }
