+++ graduacao/src/main/java/br/usp/exemplo/graduacao/servicos/MatriculaServico.java
@@ -1,33 +1,50 @@
 package br.usp.exemplo.graduacao.servicos;
 
-import br.usp.exemplo.graduacao.entidades.Matricula;
-import br.usp.exemplo.graduacao.repositorios.MatriculaRepositorio;
-import org.springframework.stereotype.Service;
-
 import java.util.List;
-import java.util.Optional;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
+
+import br.usp.exemplo.graduacao.dtos.MatriculaDTO;
+import br.usp.exemplo.graduacao.entidades.Matricula;
+import br.usp.exemplo.graduacao.forms.MatriculaForm;
+import br.usp.exemplo.graduacao.repositorios.MatriculaRepositorio;
+import lombok.RequiredArgsConstructor;
 
 @Service
+@RequiredArgsConstructor
 public class MatriculaServico {
     private final MatriculaRepositorio matriculaRepositorio;
 
-    public MatriculaServico(MatriculaRepositorio matriculaRepositorio) {
-        this.matriculaRepositorio = matriculaRepositorio;
+    public List<MatriculaDTO> listarTodos() {
+        return matriculaRepositorio.findAll().stream()
+                .map(MatriculaDTO::new)
+                .toList();
     }
 
-    public List<Matricula> listarTodos() {
-        return matriculaRepositorio.findAll();
+    public MatriculaDTO buscarPorId(Long id) {
+        return matriculaRepositorio.findById(id)
+                .map(MatriculaDTO::new)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Matrícula não encontrada com o ID: " + id));
     }
 
-    public Optional<Matricula> buscarPorId(int id) {
-        return matriculaRepositorio.findById(id);
+    public MatriculaDTO salvar(MatriculaForm matricula) {
+        Matricula novo = new Matricula(matricula);
+        return new MatriculaDTO(matriculaRepositorio.save(novo));
     }
 
-    public Matricula salvar(Matricula matricula) {
-        return matriculaRepositorio.save(matricula);
+    public MatriculaDTO atualizar(Long id, MatriculaForm form) {
+        Matricula matriculaExistente = matriculaRepositorio.findById(id)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Matrícula não encontrada com o ID: " + id));
+        matriculaExistente.atualizarDados(form);
+        return new MatriculaDTO(matriculaRepositorio.save(matriculaExistente));
     }
 
-    public void deletar(int id) {
+    public void deletar(Long id) {
+        if(!matriculaRepositorio.existsById(id)){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Matrícula não encontrada com o ID: " + id);
+        }
         matriculaRepositorio.deleteById(id);
     }
 }
